added support-ticket package. add web based installer

+++ routes/web.php
@@ -1,7 +1,12 @@
 <?php
 
+use App\Http\Controllers\RoleController;
+use App\Http\Controllers\SetupController;
+use App\Http\Controllers\TestDBController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Storage;
 
 /*
 |--------------------------------------------------------------------------
@@ -13,28 +18,50 @@ use Illuminate\Support\Facades\Route;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
+$installed = Storage::disk('public')->exists('installed');
+if ($installed === true) {
+    Route::get('/', function () {
+        return view('layouts.app');
+    });
 
-Route::get('/', function () {
-    return view('layouts.app');
-});
-
-Auth::routes(['verify' => false]);
+    Auth::routes(['verify' => false]);
 
 // Auth middleware for authenticated users
-Route::middleware(['auth'])->group(function () {
-    Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+    Route::middleware(['auth'])->group(function () {
+        Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-    Route::get('permissions', [App\Http\Controllers\PermissionController::class, 'index'])->name('permissions.index');
-    Route::get('roles/{role}/destroy', [App\Http\Controllers\RoleController::class, 'destroy'])->name('croles.destroy');
+        Route::get('permissions', [App\Http\Controllers\PermissionController::class, 'index'])->name('permissions.index');
 
-    Route::post('users/{user}', [App\Http\Controllers\UserController::class, 'destroy'])->name('users.destroy');
-    Route::post('users/{id}/force-delete', [App\Http\Controllers\UserController::class, 'forceDelete'])->name('users.forceDelete');
-    Route::get('users/{id}/retrieve', [App\Http\Controllers\UserController::class, 'retrieveDeleted'])->name('users.retrieveDeleted');
-    Route::put('users/{id}/change-password', [App\Http\Controllers\UserController::class, 'changePassword'])->name('users.changePassword');
+        Route::post('users/{id}/force-delete', [App\Http\Controllers\UserController::class, 'forceDelete'])->name('users.forceDelete');
+        Route::get('users/{id}/retrieve', [App\Http\Controllers\UserController::class, 'retrieveDeleted'])->name('users.retrieveDeleted');
+        Route::put('users/{id}/change-password', [App\Http\Controllers\UserController::class, 'changePassword'])->name('users.changePassword');
 
-    Route::resources([
-        'users' => \App\Http\Controllers\UserController::class,
-        'roles' => \App\Http\Controllers\RoleController::class,
-    ]);
-});
+        Route::resources(['users' => UserController::class, 'roles' => RoleController::class,]);
+    });
 
+} else {
+    /*
+    ------------------------------Install application-------------------------------
+    */
+    Route::get('/{url?}', function () {
+        return redirect('/setup');
+    })->where('url', '^(?!setup).*$');
+    Route::get('/setup', [SetupController::class, 'viewSetup'])->name('viewSetup');
+    Route::get('/setup/step-1', [SetupController::class, 'viewStep1'])->name('viewStep1');
+    Route::get('/setup/step-2', [SetupController::class, 'viewStep2'])->name('viewStep2');
+    Route::get('/setup/step-3', [SetupController::class, 'viewStep3'])->name('viewStep3');
+    Route::get('/setup/step-4', [SetupController::class, 'viewStep3'])->name('viewStep4');
+    Route::get('/setup/getNewAppKey', [SetupController::class, 'getNewAppKey'])->name('getNewAppKey');
+    Route::post('/setup/testDB', [TestDBController::class, 'testDB'])->name('testDB');
+    Route::post('/setup/step-2', [SetupController::class, 'setupStep1'])->name('setupStep1');
+    Route::post('/setup/step-3', [SetupController::class, 'setupStep2'])->name('setupStep2');
+    Route::post('/setup/step-4', [SetupController::class, 'setupStep3'])->name('setupStep3');
+    Route::post('/setup/lastStep', [SetupController::class, 'lastStep'])->name('lastStep');
+    Route::get('/setup/lastStep', function () {
+        return redirect('/setup', 301);
+    });
+    Route::get('/setup/finish', function () {
+
+        return view('setup.finished');
+    });
+}
